Moved json tests to json folder
