+++ telemetry/collector/src/test/java/ru/yandex/practicum/json/SensorEventJsonTest.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum;
+package ru.yandex.practicum.json;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
